freelance-pm-copilot/functions/requirements.txt
functions-framework==3.*
firebase-functions==0.1.0
firebase-admin==6.4.0
groq==0.4.1
llama-cloud-services==0.6.76
python-dotenv>=1.0.1
requests==2.31.0
